vcf/record: Move Genotypes to its own module
